\name{zigguserDirectCpp-package}
\alias{zigguserDirectCpp-package}
\alias{zigguserDirectCpp}
\docType{package}
\title{\packageTitle{zigguserDirectCpp}}
\description{\packageDescription{zigguserDirectCpp}}
\section{Package Content}{\packageIndices{zigguserDirectCpp}}
\author{\packageAuthor{zigguserDirectCpp}}
\section{Maintainer}{\packageMaintainer{zigguserDirectCpp}}
\keyword{package}
